feat!: update theme to work with 2025.8
This updates the theme to work with modern versions of Home Assistant.

BREAKING CHANGE: If you have any custom places in your dashboards where
you are using the Catppuccin css variables directly, you will need to
update them to have the `catppuccin-` prefix. For example, if you have a
custom card configured with `color: var(--rosewater)`, it needs to be
changed to `color: var(--catppuccin-rosewater)`

Fixes #7