fix: explicitly set primary color
This is a no-op for most Home Assistant pages, but it does fix color
theming in HACS. This is due to the HACS page loading an old version of
the default theme variables that used an explicit hex code instead of
the modern `var()`. It's also probably just a good idea to actually set
both of the explicitly supported theme variables, not just one of them.